allow infinite scroll
allow the user to load the next 50 results as they scroll to the bottom. As Core.ac.uk is a bit slow, I start the load at 25 records - which might be premature, but allows data to load.
A spinner at the bottom indicates that there is more data to come